Extract documentaiton generation parts
